Add to simpa and integration tests
