+++ testing/integration/Cargo.toml
@@ -8,6 +8,7 @@ include.workspace = true
 license.workspace = true
 
 [dependencies]
+kaspa-alloc.workspace = true
 kaspa-addresses.workspace = true
 kaspa-consensus-core.workspace = true
 kaspa-consensus-notify.workspace = true
@@ -53,6 +54,7 @@ smallvec.workspace = true
 tempfile.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
+ctor = "0.2.6"
 
 [dev-dependencies]
 criterion.workspace = true
