+++ utils/alloc/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "kaspa-alloc"
+description = "Kaspa allocator wrapper"
+version.workspace = true
+authors.workspace = true
+license.workspace = true
+edition.workspace = true
+include.workspace = true
+
+[dependencies]
+mimalloc.workspace = true
